+++ wondrous-app/pages/dashboard/bounties.tsx
@@ -1,35 +1,19 @@
+import React from 'react';
+
 import MetaTags from 'components/MetaTags';
-import Wrapper from 'components/Dashboard/wrapper';
-import { withAuth } from 'components/Auth/withAuth';
-import MobileComingSoonModal from 'components/Onboarding/MobileComingSoonModal';
-import { useIsMobile } from 'utils/hooks';
-import { ViewType } from 'types/common';
-import { useRouter } from 'next/router';
-import Board from 'components/Dashboard/bounties';
+import BoardSkeleton from 'components/Dashboard/boards/BoardSkeleton';
 import { getServerSideProps } from 'utils/board/dataFetching';
+import lazy from 'utils/enhancements/lazy';
 
-type Props = {
-  meta: {
-    title: string;
-    img: string;
-    description: string;
-  };
-};
+const Bounties = lazy(() => import('./index.lazy'), BoardSkeleton);
 
-const BountiesPage = ({ meta }: Props) => {
-  const isMobile = useIsMobile();
-  const router = useRouter();
-  const isAdmin = router.query.view === ViewType.Admin;
+const BountiesPage = (props) => (
+  <>
+    <MetaTags meta={props.meta} />
+    <Bounties {...props} />
+  </>
+);
 
-  return (
-    <Wrapper isAdmin={isAdmin}>
-      <MetaTags meta={meta} />
-      {isMobile ? <MobileComingSoonModal /> : null}
-      <Board isAdmin={isAdmin} />
-    </Wrapper>
-  );
-};
-
-export default withAuth(BountiesPage);
+export default BountiesPage;
 
 export { getServerSideProps };
