+++ wondrous-app/pages/dashboard/proposals.tsx
@@ -1,32 +1,19 @@
-import MetaTags from 'components/MetaTags';
-import Wrapper from 'components/Dashboard/wrapper';
-import { withAuth } from 'components/Auth/withAuth';
-import MobileComingSoonModal from 'components/Onboarding/MobileComingSoonModal';
 import React from 'react';
-import { useIsMobile } from 'utils/hooks';
-import Board from 'components/Dashboard/proposals';
+
+import MetaTags from 'components/MetaTags';
+import BoardSkeleton from 'components/Dashboard/boards/BoardSkeleton';
 import { getServerSideProps } from 'utils/board/dataFetching';
+import lazy from 'utils/enhancements/lazy';
 
-type Props = {
-  meta: {
-    title: string;
-    img: string;
-    description: string;
-  };
-};
+const Proposals = lazy(() => import('./proposals.lazy'), BoardSkeleton);
 
-const ProposalsPage = ({ meta }: Props) => {
-  const isMobile = useIsMobile();
+const ProposalsPage = (props) => (
+  <>
+    <MetaTags meta={props.meta} />
+    <Proposals {...props} />
+  </>
+);
 
-  return (
-    <Wrapper isAdmin={false}>
-      <MetaTags meta={meta} />
-      {isMobile ? <MobileComingSoonModal /> : null}
-      <Board />
-    </Wrapper>
-  );
-};
-
-export default withAuth(ProposalsPage);
+export default ProposalsPage;
 
 export { getServerSideProps };
